chore: 폴더 common 내부로 공용 컴포넌트 이동
