--- frontend/src/components/Check.tsx
@@ -1,15 +1,0 @@
-import styled from '@emotion/styled';
-
-const Button = styled.button`
-   width: 100px; 
-   height: 10px;
-  border: 1px solid red;
-
-   color: hotpink;
-`;
-
-const Check = () => {
-  return <Button />;
-};
-
-export default Check;
