+++ frontend/src/pages/ChecklistPage.tsx
@@ -1,9 +1,9 @@
 import { useCallback, useEffect, useState } from 'react';
 
 import mockCategories from '@/_mock/checklist.json';
-import Accordion from '@/components/Accordion/Accordion';
 import ChecklistCategory from '@/components/Checklist/ChecklistCategory';
-import Tabs from '@/components/Tabs/Tabs';
+import Accordion from '@/components/common/Accordion/Accordion';
+import Tabs from '@/components/common/Tabs/Tabs';
 import { ChecklistCategoryQuestions } from '@/types/checklist';
 
 export interface Answer {
