+++ frontend/src/pages/ChecklistPreviewPage.tsx
@@ -1,9 +1,9 @@
 import { useEffect, useState } from 'react';
 
 import mockCategories from '@/_mock/checklist.json';
-import Accordion from '@/components/Accordion/Accordion';
 import ChecklistCategory from '@/components/Checklist/ChecklistCategory';
-import Header from '@/components/Header/Header';
+import Accordion from '@/components/common/Accordion/Accordion';
+import Header from '@/components/common/Header/Header';
 import { ChecklistCategoryQuestions } from '@/types/checklist';
 
 export interface addAnswerProps {
