+++ frontend/src/pages/SaveCheckListPage.tsx
@@ -2,7 +2,7 @@ import styled from '@emotion/styled';
 import { Link } from 'react-router-dom';
 
 import BangGgoodIcon from '@/assets/bang-ggood-icon.svg';
-import Header from '@/components/Header/Header';
+import Header from '@/components/common/Header/Header';
 
 const SaveCheckListPage = () => {
   return (
